+++ incuna_auth/middleware/login_required.py
@@ -1,32 +1,34 @@
-import re
-
 from django.conf import settings
-from django.contrib import messages
-from django.contrib.auth.views import redirect_to_login
-from django.http import HttpResponseForbidden
-from django.utils.translation import ugettext_lazy as _
+from django.core.exceptions import ImproperlyConfigured
 
-# Python 2/3 compatibility hackery
-try:
-    unicode
-except NameError:
-    unicode = str
+from .permission import LoginPermissionMiddlewareMixin, UrlPermissionMiddleware
+from .utils import compile_urls
 
 
-LOGIN_EXEMPT_URLS = getattr(settings, 'LOGIN_EXEMPT_URLS', [])
-LOGIN_PROTECTED_URLS = getattr(settings, 'LOGIN_PROTECTED_URLS', [r'^'])
+def check_request_has_user():
+    """
+    Check that LoginRequiredMiddleware isn't being used without its dependency.
 
-EXEMPT_URLS = [
-    re.compile('^%s$' % settings.LOGIN_URL.lstrip('/')),
-    re.compile('^%s$' % settings.LOGOUT_URL.lstrip('/')),
-]
-EXEMPT_URLS += [re.compile(unicode(expr)) for expr in LOGIN_EXEMPT_URLS]
-PROTECTED_URLS = [re.compile(unicode(expr)) for expr in LOGIN_PROTECTED_URLS]
-SEND_MESSAGE = getattr(settings, 'LOGIN_REQUIRED_SEND_MESSAGE', True)
+    LoginRequiredMiddleware needs django.contrib.auth.middleware.AuthenticationMiddleware.
+    """
+    middlewares = settings.MIDDLEWARE_CLASSES
+
+    if 'django.contrib.auth.middleware.AuthenticationMiddleware' in middlewares:
+        return
+
+    error_message = ' '.join((
+        "MIDDLEWARE_CLASSES does not contain AuthenticationMiddleware.",
+        "LoginRequiredMiddleware requires authentication middleware to be",
+        "installed. Ensure that your MIDDLEWARE_CLASSES setting includes",
+        "'django.contrib.auth.middleware.AuthenticationMiddleware'.",
+    ))
+    raise ImproperlyConfigured(error_message)
 
 
-class LoginRequiredMiddleware:
-    '''
+class LoginRequiredMiddleware(LoginPermissionMiddlewareMixin, UrlPermissionMiddleware):
+    """
+    Middleware that requires a user to be authenticated.
+
     Middleware that requires a user to be authenticated to view any page in
     LOGIN_PROTECTED_URLS other than LOGIN_URL. Exemptions to this requirement
     can optionally be specified in settings via a list of regular expressions
@@ -38,40 +40,25 @@ class LoginRequiredMiddleware:
     Will default to protecting everything if LOGIN_PROTECTED_URLS is not in
     settings.
 
-
     Original code from:
     http://onecreativeblog.com/post/59051248/django-login-required-middleware
 
     This version has been modified to allow us to define areas of the site to
     password protect instead of protecting everything under /.
-    '''
-    def process_request(self, request):
-        assert hasattr(request, 'user'), (
-            "The 'LoginRequiredMiddleware' "
-            + "requires authentication middleware to be installed. Edit your "
-            + "'MIDDLEWARE_CLASSES' setting to insert "
-            + "'django.contrib.auth.middlware.AuthenticationMiddleware'. "
-            + "If that doesn't work, ensure your 'TEMPLATE_CONTEXT_PROCESSORS' "
-            + "setting includes 'django.core.context_processors.auth'."
-        )
+    """
+    login_exempt_urls = [settings.LOGIN_URL, settings.LOGOUT_URL]
+    login_exempt_urls += getattr(settings, 'LOGIN_EXEMPT_URLS', [])
+    login_protected_urls = getattr(settings, 'LOGIN_PROTECTED_URLS', [r'^'])
 
-        # Jump over this middleware if not a protected url.
-        path = request.path_info.lstrip('/')
-        path_is_exempt = any(m.match(path) for m in EXEMPT_URLS)
-        path_is_protected = any(m.match(path) for m in PROTECTED_URLS)
-        if path_is_exempt or not path_is_protected:
-            return
+    EXEMPT_URLS = compile_urls(login_exempt_urls)
+    PROTECTED_URLS = compile_urls(login_protected_urls)
 
-        # Jump over this middleware if user logged in.
-        if request.user.is_authenticated():
-            return
+    def __init__(self, check=True):
+        if check:
+            check_request_has_user()
 
-        # Raise a 403 for POST/DELETE etc.
-        if request.method != 'GET':
-            return HttpResponseForbidden()
+    def get_exempt_url_patterns(self):
+        return self.EXEMPT_URLS
 
-        # Add a message, and redirect to login.
-        if SEND_MESSAGE:
-            messages.info(request, _('You must be logged in to view this page.'))
-
-        return redirect_to_login(request.path_info)
+    def get_protected_url_patterns(self):
+        return self.PROTECTED_URLS