+++ tests/run.py
@@ -14,6 +14,8 @@ settings.configure(
     INSTALLED_APPS=(
         'incuna_auth',
 
+        'incuna_test_utils',
+
         # Put contenttypes before auth to work around test issue.
         # See: https://code.djangoproject.com/ticket/10827#comment:12
         'django.contrib.contenttypes',